tests/requirements.txt
# Test dependencies for Image2Model test suite
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-cov==4.1.0

# HTTP and async testing
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Image processing for test fixtures
Pillow==10.1.0

# YAML parsing for Docker Compose tests
PyYAML==6.0.1

# Utilities
colorama==0.4.6
tabulate==0.9.0